split the web api/web app service collection extention methods and created ones for the builder in a separate class to simplify the api update unit tests
